requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.28.0
pyyaml>=6.0.0
openpyxl>=3.1.0 